fix(review): extract projected artifact lifecycle
