+++ src/autoskillit/core/runtime/AGENTS.md
@@ -7,7 +7,7 @@ Process-state modules for session lifecycle tracking (stdlib-only, IL-0).
 | File | Purpose |
 |------|---------|
 | `__init__.py` | Named explicit re-exports (not wildcard) for all public symbols |
-| `artifact_lease.py` | Stable regular-file shared/exclusive leases inherited by artifact-consuming children |
+| `artifact_lease.py` | Stable regular-file leases plus shared error-preserving launch-binding scope |
 | `_linux_proc.py` | Reads `/proc` boot ID and process start-time ticks; returns `None` on non-Linux |
 | `kitchen_state.py` | `KitchenMarker` disk-based session marker written by `open_kitchen`, read by hook subprocesses; also `find_caller_session_id()` |
 | `readiness.py` | Filesystem sentinel for MCP server startup synchronization in integration tests |
